fix: fixed eslint errors
