+++ src/components/Header/Header.jsx
@@ -6,9 +6,9 @@ import {
   Form,
   FormControl,
   Button,
-} from "react-bootstrap";
+} from 'react-bootstrap';
 
-import "./header.scss"
+import './header.scss';
 
 function Header() {
   return (
@@ -20,7 +20,7 @@ function Header() {
           <Navbar.Collapse id="navbarScroll">
             <Nav
               className="me-auto my-2 my-lg-0"
-              style={{ maxHeight: "100px" }}
+              style={{ maxHeight: '100px' }}
               navbarScroll
             >
               <Nav.Link href="#action1">Home</Nav.Link>
