+++ src/components/Pages/index.js
@@ -1,4 +1,4 @@
-import MainPage from "./MainPage";
-import AnotherPage from "./AnotherPage";
+import MainPage from './MainPage';
+import AnotherPage from './AnotherPage';
 
 export { MainPage, AnotherPage };
