--- src/index.js
@@ -1,16 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App/App";
-import { BrowserRouter as Router } from "react-router-dom";
-
-import "./components/style/style.scss";
-import "bootstrap/dist/css/bootstrap.min.css";
-
-ReactDOM.render(
-  <React.StrictMode>
-    <Router>
-      <App />
-    </Router>
-  </React.StrictMode>,
-  document.getElementById("root")
-);
